added attribute painter, and export space attributes
You can now paint attributes including new paint materials that get
converted to stub space types, thermal zones, or building stories when
merged back to OSM. Merge to OSM is slow, part of that is logic to
inspect geometry, but also a lot of bad code that can be cleaned up, but
not a priority for now. Save to SKP is almost instant.